% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lq-dsl.R
\name{lq-dsl}
\alias{lq-dsl}
\title{Higher level query tools}
\description{
Higher level query tools
}
\examples{
(x <- list(a = list(b = list(c = 1, d = 2, e = 3)), f = 45))

library(jqr)
(tj(x) \%>\% index()) \%>\% fj
(tj(x) \%>\% index() \%>\% index()) \%>\% fj
(tj(x) \%>\% index() \%>\% index() \%>\% select(. > 2)) \%>\% fj
(tj(x) \%>\% index() \%>\% select(. > 2)) \%>\% fj
(tj(x) \%>\% index() \%>\% index() \%>\% keys() \%>\% reverse) \%>\% fj
}
