% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lq.R
\name{lq}
\alias{lq}
\title{Low level query tool}
\usage{
lq(x, program, ...)
}
\arguments{
\item{x}{a list}

\item{program}{a \code{jqr} query, see \code{\link[jqr:jq]{jqr::jq()}}}

\item{...}{args passed on to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}}
}
\description{
Low level query tool
}
\examples{
x <- list(a = list(b = list(c = 1:10, d = 1:10)))
lq(x, ".")
lq(x, ".[]")
lq(x, ".[][]")
lq(x, ".[][] | keys | reverse")
}
